# docs/swagger.yaml
basePath: /
definitions:
...
  create.Request:
    properties:
      barcode:
        type: integer
      category_id:
        type: string
      message:
        type: string
    required:
    - category_id
    - message
    type: object
  domain.Category:
    properties:
      answer:
        type: string
      description:
        description: Detailed description of the categories
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  domain.Complaint:
    properties:
      answer:
        example: Complaint resolved
        type: string
      barcode:
        example: 242590
        type: integer
      category:
        $ref: '#/definitions/domain.Category'
      created_at:
        example: "2025-04-21T12:00:00Z"
        type: string
      id:
        example: 71cb6c7d-5c1c-4a5f-bc08-204b6d435c25
        type: string
      message:
        example: Too noisy in the dorm
        type: string
      status:
        example: approved
        type: string
      updated_at:
        example: "2025-04-21T14:00:00Z"
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
    required:
    - statusCode
    type: object
  update.Request:
    properties:
      data:
        $ref: '#/definitions/domain.Complaint'
    type: object
...
host: complaints-api.yeunikey.dev
info:
  contact:
    email: quanaimadi@.gmail.com
    name: API Support
  description: This is a server for managing complaints and categories.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Complaint Server API
  version: "1.0"
paths:
...
  /admin/complaints/{id}:
    delete:
      description: Delete a complaint by its ID. If the complaint is not found, an
        error is returned.
      parameters:
      - description: Complaint ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Complaint successfully deleted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request or complaint not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      tags:
      - Complaints
    put:
      consumes:
      - application/json
      description: Updates an existing complaint based on the provided complaint ID
        and new data.
      parameters:
      - description: Complaint ID
        in: path
        name: id
        required: true
        type: string
      - description: Complaint resolution details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/update.Request'
      produces:
      - application/json
      responses:
        "200":
          description: Complaint updated successfully
          schema:
            $ref: '#/definitions/update.Request'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Complaint not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      tags:
      - Complaints
...
  /complaints:
    get:
      consumes:
      - application/json
      description: Retrieve a complaint using its unique identifier. The ID must be
        an integer that corresponds to a valid complaint in the database.
      parameters:
      - description: Complaint ID (unique identifier of the complaint)
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Complaint details
          schema:
            $ref: '#/definitions/domain.Complaint'
        "400":
          description: Invalid request, incorrect ID format
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Complaint with the given ID not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get a complaint by ID
      tags:
      - Complaints
    post:
      consumes:
      - application/json
      description: Create a new complaint for a specific user and categories. Only one
        complaint can be submitted per hour for the same user.
      parameters:
      - description: Complaint details
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/create.Request'
      produces:
      - application/json
      responses:
        "200":
          description: Success response with complaint ID and answer
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request, bad input or validation error
          schema:
            $ref: '#/definitions/response.Response'
        "429":
          description: Limit of one complaint per hour exceeded
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create a new complaint
      tags:
      - Complaints
  /complaints/{id}:
    delete:
      description: Delete a complaint by its ID by owner. If the complaint is not
        found or user is not owner, an error is returned.
      parameters:
      - description: Complaint ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Complaint successfully deleted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request or complaint not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      tags:
      - Complaints
    get:
      consumes:
      - application/json
      description: Retrieve a complaint using its unique identifier. The UUID must
        be a string that corresponds to a valid complaint in the database.
      parameters:
      - description: Complaint ID (unique identifier of the complaint)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Complaint details
          schema:
            $ref: '#/definitions/domain.Complaint'
        "400":
          description: Invalid request, incorrect ID format
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Complaint with the given ID not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get a complaint by ID
      tags:
      - Complaints
  /complaints/by-token:
    get:
      consumes:
      - application/json
      description: Retrieves all complaints associated with a user based on the provided
        token.
      parameters:
      - description: User token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of complaints
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Token required
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Invalid token or failed to fetch profile
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to serialize complaints
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get user complaints by token
      tags:
      - Complaints
  /complaints/can-submit:
    get:
      consumes:
      - application/json
      description: Проверяет, может ли пользователь отправить новую жалобу (прошел
        ли час с последнего запроса)
      parameters:
      - description: Barcode пользователя
        in: query
        name: barcode
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: true/false
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: missing barcode
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Проверка возможности отправки жалобы
      tags:
      - Complaints
schemes:
- https
swagger: "2.0"